mobile: fix unable to edit/delete reminder when accessing reminders from note properties
